Keep moving from old ocfile to new one 5
